MVP Data Capture Stage 1
This first commit includes only what's required for data capture, while the training modules of the platform are being worked on. During training, we will develop the prediction components.

This first update includes a rough first draft of:
1. Timestamped Keystroke logging (Please use with caution!), for capturing typing intent.
2. Timestamped EEG data capture, currently only supporting Neurosity Crown.

This is the first seed of what Singularity might one day look like, at this stage we are trying to find a viable model for Thought to Text prediction at a comparable speed (words per minute) close to the average humans typing.